상세페이지 댓글 수정, 신고 기능 추가

+++ src/main/java/project/spring/nft/controller/ArtController.java
@@ -3,11 +3,15 @@ package project.spring.nft.controller;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.PrintWriter;
 import java.util.List;
 import java.util.Map;
 
 import javax.annotation.Resource;
+import javax.mail.internet.MimeMessage;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
@@ -16,6 +20,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,6 +34,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import project.spring.nft.domain.ArtVO;
 import project.spring.nft.domain.AuctionVO;
+import project.spring.nft.domain.MemberVO;
+import project.spring.nft.domain.PaymentVO;
 import project.spring.nft.pageutil.PageCriteria;
 import project.spring.nft.pageutil.PageMaker;
 import project.spring.nft.service.ArtService;
@@ -46,6 +54,9 @@ public class ArtController {
 	
 	@Autowired
 	private ArtService artService;
+	
+	@Autowired
+	private JavaMailSender mailSender;
 	
 	@GetMapping("main")
 	public void main(Model model, Integer page, Integer numsPerPage) {
@@ -83,6 +94,7 @@ public class ArtController {
 		pageMaker.setPageData();
 		logger.info("이전 버튼 존재 유무 : "+pageMaker.isHasPrev());
 		logger.info("다음 버튼 존재 유무 : "+pageMaker.isHasNext());
+		logger.info("전체 게시글 수 : "+pageMaker.getTotalCount());
 		model.addAttribute("pageMaker", pageMaker);
 	} //end currentAllList()
 	
@@ -106,6 +118,9 @@ public class ArtController {
 		pageMaker.setCriteria(criteria);
 		pageMaker.setTotalCount(artService.getTotalNumsOfRecords());
 		pageMaker.setPageData();
+		logger.info("이전 버튼 존재 유무 : "+pageMaker.isHasPrev());
+		logger.info("다음 버튼 존재 유무 : "+pageMaker.isHasNext());
+		logger.info("전체 게시글 수 : "+pageMaker.getTotalCount());
 		model.addAttribute("pageMaker", pageMaker);
 		
 		return "main";
@@ -131,6 +146,9 @@ public class ArtController {
 		pageMaker.setCriteria(criteria);
 		pageMaker.setTotalCount(artService.getTotalNumsOfRecords());
 		pageMaker.setPageData();
+		logger.info("이전 버튼 존재 유무 : "+pageMaker.isHasPrev());
+		logger.info("다음 버튼 존재 유무 : "+pageMaker.isHasNext());
+		logger.info("전체 게시글 수 : "+pageMaker.getTotalCount());
 		model.addAttribute("pageMaker", pageMaker);
 		
 		return "main";
@@ -167,6 +185,9 @@ public class ArtController {
 
 		pageMaker.setCriteria(criteria);
 		pageMaker.setPageData();
+		logger.info("이전 버튼 존재 유무 : "+pageMaker.isHasPrev());
+		logger.info("다음 버튼 존재 유무 : "+pageMaker.isHasNext());
+		logger.info("전체 게시글 수 : "+pageMaker.getTotalCount());
 		model.addAttribute("pageMaker", pageMaker);
 		
 		return "main";
@@ -263,16 +284,35 @@ public class ArtController {
 			int maxMoney=(Integer)readMap.get("maxMoney");
 			model.addAttribute("maxMoney", maxMoney);			
 		}
+		
+		PaymentVO pvo=artService.readPayResult(artNo);
+		String payResult="";
+		if(pvo != null) { //결제된 작품이다
+			payResult="fail";
+		}else {
+			payResult="success";
+		}
+		
+		model.addAttribute("payResult", payResult);
 		model.addAttribute("vo", vo);
 		model.addAttribute("page", page);
 	} //end detail()
 	
 	@GetMapping("/arts/update")
-	public void updateGET(Model model, Integer artNo) {
+	public void updateGET(Model model, Integer artNo, HttpServletRequest request) {
 		logger.info("updateGET() 호출 : artNo = "+artNo);
+		
+		HttpSession session = request.getSession();
+		String memberId = (String) session.getAttribute("memberId");
+		
 		Map<String, Object> readMap=artService.readArtNo(artNo);
 		ArtVO vo=(ArtVO)readMap.get("vo");
-		model.addAttribute("vo", vo);
+		if(memberId.equals(vo.getMemberId())) {
+			model.addAttribute("vo", vo);
+			model.addAttribute("access", "success");
+		}else {
+			model.addAttribute("access", "fail");
+		}
 	} //end updateGET()
 	
 	@PostMapping("arts/update")
@@ -290,16 +330,26 @@ public class ArtController {
 	} //end updatePOST()
 
 	@GetMapping("arts/delete")
-	public String deletePOST(int artNo, RedirectAttributes reAttr) throws Exception {
+	public String deletePOST(int artNo, RedirectAttributes reAttr,
+			HttpServletRequest request) throws Exception {
 		logger.info("deletePOST() 호출");
-		int result=artService.deleteArt(artNo);
+		HttpSession session = request.getSession();
+		String memberId = (String) session.getAttribute("memberId");
 		
-		if(result==1) {
-			reAttr.addFlashAttribute("deleteResult", "success"); 
-			return "redirect:/main"; 
-		}else {
+		ArtVO vo=artService.readArtno(artNo);
+		if(memberId.equals(vo.getMemberId())) {
+			int result=artService.deleteArt(artNo);
+			
+			if(result==1) {
+				reAttr.addFlashAttribute("deleteResult", "success"); 
+				return "redirect:/main"; 
+			}else {
+				reAttr.addFlashAttribute("deleteResult", "fail"); 
+				return "redirect:/arts/detail?artNo="+artNo;
+			}			
+		}else { //작성자가 아닌 경우
 			reAttr.addFlashAttribute("deleteResult", "fail"); 
-			return "redirect:/arts/detail?artNo="+artNo;
+			return "redirect:/main";
 		}
 	} //end deletePOST()
 	
@@ -317,4 +367,58 @@ public class ArtController {
 			return bidList;			
 		}
 	} //end winningBid()
+	
+	@GetMapping("arts/declare")
+	public void declareGET(int artNo, HttpServletRequest request, Model model) {
+		logger.info("declareGET() 호출");
+		HttpSession session = request.getSession();
+		String memberId = (String) session.getAttribute("memberId");
+		/*
+		 * if(memberId == null) { model.addAttribute("accessResult", "fail");
+		 * logger.info(memberId); } else { model.addAttribute("accessResult",
+		 * "success"); logger.info(memberId); }
+		 */
+		model.addAttribute("artNo", artNo);
+		MemberVO vo = artService.readMemberEmail(memberId);
+		logger.info(vo.toString());
+		String memberEmail = vo.getMemberEmail();
+		model.addAttribute("memberEmail", memberEmail);
+	}
+	
+	@PostMapping("arts/declare")
+	public void declarePOST(HttpServletResponse response, RedirectAttributes reAttr, String memberEmail, int artNo, String declareContent) throws Exception {
+		logger.info("declarePOST() 호출");
+		int result = sendMail(memberEmail, artNo, declareContent);
+		if(result == 1) {
+			logger.info(result+"개 메일 발송 완료");
+			PrintWriter out = response.getWriter();
+			out.println("<script> alert('신고 접수가 완료되었습니다.'); window.close(); </script>");
+			reAttr.addFlashAttribute("emailResult", "success");
+		} else {
+			logger.info("메일 발송 실패");
+			PrintWriter out = response.getWriter();
+			out.println("<script> alert('다시 시도해주세요'); </script>");
+			reAttr.addFlashAttribute("emailResult", "fail");
+		}
+	}
+	
+	// 작품 신고
+	public int sendMail(String memberEmail, int artNo, String declareContent) throws Exception {
+		int result = 0;
+		String subject = artNo + "번 작품 신고 접수";
+		String content = "신고 내용 : " + declareContent;
+		String from = memberEmail;
+		String to = "hansl2249@naver.com";
+		
+		MimeMessage mail = mailSender.createMimeMessage();
+		MimeMessageHelper mailHelper = new MimeMessageHelper(mail, "UTF-8");
+		mailHelper.setFrom(from);
+		mailHelper.setTo(to);
+		mailHelper.setSubject(subject);
+		mailHelper.setText(content, true);
+		
+		mailSender.send(mail);
+		result = 1;
+		return result;
+	} // end sendMail()
 }
