+++ src/main/java/project/spring/nft/service/ArtServiceImple.java
@@ -1,6 +1,8 @@
 package project.spring.nft.service;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -11,10 +13,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import project.spring.nft.domain.ArtVO;
+import project.spring.nft.domain.MemberVO;
+import project.spring.nft.domain.PaymentVO;
 import project.spring.nft.pageutil.PageCriteria;
 import project.spring.nft.persistence.ArtDAO;
 import project.spring.nft.persistence.ArtReplyDAO;
 import project.spring.nft.persistence.AuctionDAO;
+import project.spring.nft.persistence.MemberDAO;
+import project.spring.nft.persistence.PaymentDAO;
 import project.spring.nft.persistence.WishlistDAO;
 
 @Service
@@ -30,6 +36,10 @@ public class ArtServiceImple implements ArtService {
 	private ArtReplyDAO artReplyDAO;
 	@Autowired
 	private WishlistDAO wishlistDAO;
+	@Autowired
+	private PaymentDAO paymentDAO;
+	@Autowired
+	private MemberDAO memberDAO;
 	
 	@Override
 	public int createArt(ArtVO vo) {
@@ -87,7 +97,7 @@ public class ArtServiceImple implements ArtService {
 
 	@Override
 	public List<ArtVO> readMemberNickname(PageCriteria criteria, String keyword) {
-		logger.info("readMemberNo() : keyword = "+keyword);
+		logger.info("readMemberNickname() : keyword = "+keyword);
 		return artDAO.selectMemberNickname(criteria, keyword);
 	}
 	
@@ -117,9 +127,9 @@ public class ArtServiceImple implements ArtService {
 	}
 
 	@Override
-	public List<ArtVO> readByMemberId(String memberId) {
+	public List<ArtVO> readByMemberId(PageCriteria criteria, String memberId) {
 		logger.info("readByMemberId() 호출 : memberId = " + memberId);
-		return artDAO.selectMemberId(memberId);
+		return artDAO.selectMemberId(criteria, memberId);
 	}
 
 	@Override
@@ -127,7 +137,7 @@ public class ArtServiceImple implements ArtService {
 		logger.info("updateWishCount() 호출 : artNo = "+artNo+", count = "+count);
 		return artDAO.updateWishCount(artNo, count);
 	}
-
+	
 	@Transactional
 	@Override
 	public int updateArt(ArtVO vo) {
@@ -141,30 +151,52 @@ public class ArtServiceImple implements ArtService {
 		return result;
 	}
 	
-
 	//작품 삭제 시 작품에 달린 댓글, 경매기록, 찜도 전부 지워져야함
 	//TODO 만약 누군가 구매했다면 작품삭제 불가능(구매기능 만들고 업데이트)
 	@Transactional
 	@Override
 	public int deleteArt(int artNo) throws Exception{
 		logger.info("deleteArt() 호출");
-		int result=artDAO.deleteArt(artNo);
-		logger.info("작품 삭제 성공");
-		
-		artReplyDAO.deleteArtNo(artNo); //작품의 댓글들 삭제
-		logger.info("작품의 댓글 삭제 성공");
-		//작품의 경매내역 삭제
-		auctionDAO.deleteArtNo(artNo);
-		logger.info("작품의 경매내역 삭제 성공");
-		//TODO 작품의 위시리스트 삭제
-		
+		int result=0;
+		if(paymentDAO.selectByArtNo(artNo) == null) {
+			result=artDAO.deleteArt(artNo);
+			logger.info("작품 삭제 성공");
+			
+			artReplyDAO.deleteArtNo(artNo); //작품의 댓글들 삭제
+			logger.info("작품의 댓글 삭제 성공");
+			//작품의 경매내역 삭제
+			auctionDAO.deleteArtNo(artNo);
+			logger.info("작품의 경매내역 삭제 성공");
+			//작품의 위시리스트 삭제
+			wishlistDAO.deleteArt(artNo);
+			logger.info("작품의 찜내역 삭제 성공");
+		}else { //해당 작품의 결제내역이 존재함
+			logger.info("결제된 작품은 삭제 불가능");
+		}
 		return result;
 	}	
 	
 	@Override
 	public List<ArtVO> readWinBid(String memberId) {
 		logger.info("readWinBid() 호출");
-		return artDAO.selectWinBid(memberId);
+		List<ArtVO> alist=artDAO.selectWinBid(memberId);
+		List<PaymentVO> plist=paymentDAO.selectAll(memberId);
+		
+		//낙찰된 작품번호랑 일치하는 구매된 작품번호가 있으면 해당 vo는 빼고 list에 넣기
+		for(int i=0;i<plist.size();i++) { //3, 4
+			for(int j=0;j<alist.size();j++) { //2, 3, 5, 7
+				if(plist.get(i).getArtNo() == alist.get(j).getArtNo()) {
+					//일치하는 artNo가 없음
+					alist.remove(alist.get(j));
+				}
+			}
+		}
+		
+		for (ArtVO vo : alist) {
+			System.out.println("alist = "+vo.toString());
+		}
+		
+		return alist;
 	}
 
 	@Override
@@ -172,12 +204,25 @@ public class ArtServiceImple implements ArtService {
 		logger.info("readArtNo() 호출");
 		return artDAO.selectArtNo(artNo);
 	}
-
+	
 	@Override
-	public List<ArtVO> readNotAuction() {
-		logger.info("readNotAuction() 호출");
-		return artDAO.selectNotAuction();
+	public PaymentVO readPayResult(int artNo) {
+		logger.info("readPayResult() 호출");
+		return paymentDAO.selectByArtNo(artNo);
+	}
+	
+	@Override
+	public int getTotalMyArt(String memberId) {
+		logger.info("getTotalMyArt() 호출");
+		return artDAO.getTotalMyArt(memberId);
 	}
 
-
+	@Override
+	public MemberVO readMemberEmail(String memberId) {
+		logger.info("readMemberEmail() 호출");
+		return memberDAO.selectByMemberId(memberId);
+	}
+	
+	
+	
 }
